backend/requirements/base.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.7
django-pgvector==0.2.1

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# Workflow Engine
django-river==3.2.0

# File Processing
python-docx-template==0.17.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.0.1

# Background Tasks & Cache
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# API Documentation
drf-spectacular==0.26.5

# Development Tools
django-debug-toolbar==4.2.0
ipython==8.16.1

# Monitoring & Logging
sentry-sdk[django]==1.35.0

# Environment Management
python-decouple==3.8